+++ netbox_pdu_control/views.py
@@ -19,7 +19,7 @@ from . import filtersets, forms, jobs, models, tables
 from .backends import _VENDOR_BACKENDS, get_pdu_client
 from .backends.base import PDUClientError
 from .choices import OutletStatusChoices, SyncStatusChoices
-from .config_backup import save_config_backup
+from .config_backup import get_config_diff, save_config_backup
 from .credentials import SECRET_ROLE_SLUG, get_credential
 from .jobs import epoch_to_dt, fetch_pdu_metrics, sync_managed_pdu
 
@@ -76,6 +76,7 @@ class ManagedPDUView(generic.ObjectView):
             "ocps": ocps,
             "secrets_available": secrets_available,
             "secret_found": secret_found,
+            "config_diff": get_config_diff(instance),
         }